+++ Makefile
@@ -6,13 +6,14 @@ CONTAINER_ENGINE ?= docker
 TARGET=hubble
 GIT_BRANCH != which git >/dev/null 2>&1 && git rev-parse --abbrev-ref HEAD
 GIT_HASH != which git >/dev/null 2>&1 && git rev-parse --short HEAD
+GOTAGS ?= hubble_serve
 
 TEST_TIMEOUT ?= 5s
 
 all: hubble
 
 hubble:
-	$(GO) build -ldflags "-X 'github.com/cilium/hubble/pkg.GitBranch=${GIT_BRANCH}' -X 'github.com/cilium/hubble/pkg.GitHash=$(GIT_HASH)'" -o $(TARGET)
+	$(GO) build $(if $(GOTAGS),-tags $(GOTAGS)) -ldflags "-X 'github.com/cilium/hubble/pkg.GitBranch=${GIT_BRANCH}' -X 'github.com/cilium/hubble/pkg.GitHash=$(GIT_HASH)'" -o $(TARGET)
 
 install:
 	groupadd -f hubble
