+++ cmd/root.go
@@ -18,24 +18,17 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"runtime"
-	"runtime/pprof"
 
 	"github.com/cilium/hubble/cmd/observe"
-	"github.com/cilium/hubble/cmd/serve"
 	"github.com/cilium/hubble/cmd/status"
 	"github.com/cilium/hubble/cmd/version"
 	"github.com/cilium/hubble/pkg"
-	"github.com/cilium/hubble/pkg/logger"
-
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
 var (
-	cfgFile                        string
-	cpuprofile, memprofile         string
-	cpuprofileFile, memprofileFile *os.File
+	cfgFile string
 )
 
 // rootCmd represents the base command when called without any subcommands
@@ -46,45 +39,6 @@ var rootCmd = &cobra.Command{
 	SilenceErrors: true, // this is being handled in main, no need to duplicate error messages
 	SilenceUsage:  true,
 	Version:       pkg.Version,
-	PersistentPreRunE: func(_ *cobra.Command, _ []string) error {
-		return pprofInit()
-	},
-	PersistentPostRunE: func(_ *cobra.Command, _ []string) error {
-		return pprofTearDown()
-	},
-}
-
-func pprofInit() error {
-	var err error
-	if cpuprofile != "" {
-		cpuprofileFile, err = os.Create(cpuprofile)
-		if err != nil {
-			return fmt.Errorf("failed to create cpu profile: %v", err)
-		}
-		pprof.StartCPUProfile(cpuprofileFile)
-	}
-	if memprofile != "" {
-		memprofileFile, err = os.Create(memprofile)
-		if err != nil {
-			return fmt.Errorf("failed to create memory profile: %v", err)
-		}
-	}
-	return nil
-}
-
-func pprofTearDown() error {
-	if cpuprofileFile != nil {
-		pprof.StopCPUProfile()
-		cpuprofileFile.Close()
-	}
-	if memprofileFile != nil {
-		runtime.GC() // get up-to-date statistics
-		if err := pprof.WriteHeapProfile(memprofileFile); err != nil {
-			return fmt.Errorf("failed to write memory profile: %v", err)
-		}
-		memprofileFile.Close()
-	}
-	return nil
 }
 
 // Execute adds all child commands to the root command sets flags
@@ -103,22 +57,10 @@ func init() {
 	rootCmd.AddCommand(newCmdCompletion(os.Stdout))
 	rootCmd.SetErr(os.Stderr)
 
-	rootCmd.PersistentFlags().StringVar(&cpuprofile,
-		"cpuprofile", "", "Enable CPU profiling",
-	)
-	rootCmd.PersistentFlags().StringVar(&memprofile,
-		"memprofile", "", "Enable memory profiling",
-	)
-	rootCmd.PersistentFlags().Lookup("cpuprofile").Hidden = true
-	rootCmd.PersistentFlags().Lookup("memprofile").Hidden = true
-
 	rootCmd.SetVersionTemplate("{{with .Name}}{{printf \"%s \" .}}{{end}}{{printf \"v%s\" .Version}}\n")
-
-	l := logger.GetLogger()
 
 	// initialize all subcommands
 	rootCmd.AddCommand(observe.New())
-	rootCmd.AddCommand(serve.New(l))
 	rootCmd.AddCommand(status.New())
 	rootCmd.AddCommand(version.New())
 }
